API Opportunity - Vaga [POST]

+++ backend/src/main/java/br/com/house/digital/projetointegrador/repository/ProfileRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.house.digital.projetointegrador.model.Profile;
 
 @Repository
-public interface ProfileRepository extends JpaRepository<Profile, Integer> {    
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
 
 }
